Refactor a filter function to be reusuable. Fix the issue that the filter let a completed task into the list.
