+++ lib/clickEvents.js
@@ -1,6 +1,6 @@
 const $ = require("jquery");
 import * as globalVars from "./globalVars";
-import {inputCheck, mainFunction, getIncompleteTasks, sortArray} from "./miscFunctions";
+import {inputCheck, mainFunction, getIncompleteTasks, sortArray, filteredArrayList} from "./miscFunctions";
 import NewIdea, {clearTasks, showCompletedTasks, loadTopTenTasks, displayTasks, showMoreTasks} from "./taskLoader";
 import {storageArray} from "./storage";
 import {updateVote} from "./vote";
@@ -133,38 +133,21 @@ $("#show-more-button").on("click", () => {
 })
 
 $(".filter-buttons-container").on("click", (e) => {
-  const checkboxCheck = $(".filter-buttons-container").children("label");
+  const arrayToDisplay = filteredArrayList();
   let incompleteTasks = getIncompleteTasks(storageArray());
-  let numCompletedTasks = clearTasks();
-  let filterArray = [];
-  let arrayToDisplay = [];
 
   $(".task-box").remove();
   globalVars.completedDisplayed = false;
 
-  checkboxCheck.each((e) => {
-    if($(checkboxCheck[e]).children().is(":checked")){
-      filterArray.push($(checkboxCheck[e]).text());
-    }
-  })
-
-  filterArray.forEach((e) => {
-    incompleteTasks.forEach((i) => {
-      if(e === i.importance){
-        arrayToDisplay.push(i)
-      }
-    })
-  })
-
-  if( filterArray.length === 0 && globalVars.topTen && globalVars.completedDisplayed) {
-    loadTopTenTasks(sortArray(incompleteTasks), numCompletedTasks);
+  if(arrayToDisplay.length === 0 && globalVars.topTen && globalVars.completedDisplayed) {
+    loadTopTenTasks(sortArray(incompleteTasks), clearTasks());
     showCompletedTasks();
-  } else if(filterArray.length === 0 && !globalVars.topTen && globalVars.completedDisplayed) {
+  } else if(arrayToDisplay.length === 0 && !globalVars.topTen && globalVars.completedDisplayed) {
     displayTasks(sortArray(incompleteTasks));
     showCompletedTasks();
-  } else if(filterArray.length === 0 && globalVars.topTen && !globalVars.completedDisplayed) {
+  } else if(arrayToDisplay.length === 0 && globalVars.topTen && !globalVars.completedDisplayed) {
     loadTopTenTasks(sortArray(incompleteTasks));
-  } else if(filterArray.length === 0 && !globalVars.topTen && !globalVars.completedDisplayed) {
+  } else if(arrayToDisplay.length === 0 && !globalVars.topTen && !globalVars.completedDisplayed) {
     displayTasks(sortArray(incompleteTasks));
   } else {
     if(globalVars.topTen) {
@@ -173,16 +156,4 @@ $(".filter-buttons-container").on("click", (e) => {
       displayTasks(sortArray(arrayToDisplay));
     }
   }
-
-  // if(filterArray.length === 0 && topTen && !globalVars.completedDisplayed){
-  //   loadTopTenTasks(sortArray(incompleteTasks));
-  // } else if (filterArray.length === 0 && !topTen && !globalVars.completedDisplayed){
-  //   displayTasks(sortArray(incompleteTasks));
-  // } else {
-  //   if(topTen){
-  //     loadTopTenTasks(sortArray(arrayToDisplay));
-  //   } else {
-  //     displayTasks(sortArray(arrayToDisplay));
-  //   }
-  // }
 })
