+++ lib/miscFunctions.js
@@ -1,5 +1,6 @@
 const $ = require("jquery");
 import * as globalVars from "./globalVars";
+import {storageArray} from "./storage";
 import {newTask} from "./taskLoader";
 
 export const mainFunction = (obj) => {
@@ -8,7 +9,7 @@ export const mainFunction = (obj) => {
   clearFields();
 }
 
-export const displayedCheck = () => {
+export const displayedIdCheck = () => {
   let displayedTasksArray = $(".task-box");
   let displayedTasksIds = [];
 
@@ -19,6 +20,34 @@ export const displayedCheck = () => {
   });
 
   return displayedTasksIds;
+}
+
+export const filterCheckboxList = () => {
+  const checkboxCheck = $(".filter-buttons-container").children("label");
+  let filterArray = [];
+
+  checkboxCheck.each((e) => {
+    if($(checkboxCheck[e]).children().is(":checked")){
+      filterArray.push($(checkboxCheck[e]).text());
+    }
+  })
+  return filterArray;
+}
+
+export const filteredArrayList = () => {
+  const filterArray = filterCheckboxList();
+  const storedObjArray = storageArray();
+  let arrayToDisplay = [];
+
+  filterArray.forEach((e) => {
+    storedObjArray.forEach((i) => {
+      if(e === i.importance && !i.completed){
+        arrayToDisplay.push(i)
+      }
+    })
+  })
+
+  return arrayToDisplay;
 }
 
 export const inputCheck = () => {
