+++ lib/taskLoader.js
@@ -1,11 +1,9 @@
 const $ = require("jquery");
 import * as globalVars from "./globalVars";
-import {sortArray, displayedCheck, getIncompleteTasks} from "./miscFunctions";
+import {sortArray, displayedIdCheck, getIncompleteTasks, filteredArrayList} from "./miscFunctions";
 import "./clickEvents";
 import {complete} from "./completeTask";
 import {storageArray} from "./storage";
-
-
 
 export default class NewIdea {
   constructor(id, title, body, importance = "Normal", completed = false) {
@@ -68,7 +66,8 @@ export const displayTasks = (displayArray) => {
 }
 
 export const showCompletedTasks = () => {
-  let displayedTasksIds = displayedCheck();
+  const displayedTasksIds = displayedIdCheck();
+  const arrayToDisplay = filteredArrayList();
 
   for (let i = 0; i < localStorage.length; i++) {
     let storedObj = JSON.parse(localStorage.getItem(localStorage.key(i)));
@@ -88,34 +87,19 @@ export const showCompletedTasks = () => {
 export const showMoreTasks = () => {
   const checkboxCheck = $(".filter-buttons-container").children("label");
   const storedObjArray = storageArray();
+  const arrayToDisplay = filteredArrayList();
   let numCompletedTasks = clearTasks();
   let incompleteTasks = getIncompleteTasks(storedObjArray);
-  let filterArray = [];
-  let arrayToDisplay = [];
 
-  checkboxCheck.each((e) => {
-    if ($(checkboxCheck[e]).children().is(":checked")) {
-      filterArray.push($(checkboxCheck[e]).text());
-    }
-  })
-
-  filterArray.forEach((e) => {
-    storedObjArray.forEach((i) => {
-      if(e === i.importance){
-        arrayToDisplay.push(i)
-      }
-    })
-  })
-
-  if(filterArray.length === 0 && globalVars.topTen && globalVars.completedDisplayed) {
+  if(arrayToDisplay.length === 0 && globalVars.topTen && globalVars.completedDisplayed) {
     loadTopTenTasks(sortArray(incompleteTasks), numCompletedTasks);
     showCompletedTasks();
-  } else if(filterArray.length === 0 && !globalVars.topTen && globalVars.completedDisplayed) {
+  } else if(arrayToDisplay.length === 0 && !globalVars.topTen && globalVars.completedDisplayed) {
     displayTasks(sortArray(incompleteTasks));
     showCompletedTasks();
-  } else if(filterArray.length === 0 && globalVars.topTen && !globalVars.completedDisplayed) {
+  } else if(arrayToDisplay.length === 0 && globalVars.topTen && !globalVars.completedDisplayed) {
     loadTopTenTasks(sortArray(incompleteTasks));
-  } else if(filterArray.length === 0 && !globalVars.topTen && !globalVars.completedDisplayed) {
+  } else if(arrayToDisplay.length === 0 && !globalVars.topTen && !globalVars.completedDisplayed) {
     displayTasks(sortArray(incompleteTasks));
   } else {
     if(topTen) {
